Revert "feat: add io panel with json import"
